// frontend/generated/models/ProjectInCategoryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProjectInCategoryDto
 */
export interface ProjectInCategoryDto {
    /**
     * 
     * @type {number}
     * @memberof ProjectInCategoryDto
     */
    categoryId?: number;
    /**
     * 
     * @type {number}
     * @memberof ProjectInCategoryDto
     */
    success?: number;
    /**
     * 
     * @type {number}
     * @memberof ProjectInCategoryDto
     */
    total?: number;
}

/**
 * Check if a given object implements the ProjectInCategoryDto interface.
 */
export function instanceOfProjectInCategoryDto(value: object): value is ProjectInCategoryDto {
    return true;
}

export function ProjectInCategoryDtoFromJSON(json: any): ProjectInCategoryDto {
    return ProjectInCategoryDtoFromJSONTyped(json, false);
}

export function ProjectInCategoryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectInCategoryDto {
    if (json == null) {
        return json;
    }
    return {
        
        'categoryId': json['categoryId'] == null ? undefined : json['categoryId'],
        'success': json['success'] == null ? undefined : json['success'],
        'total': json['total'] == null ? undefined : json['total'],
    };
}

export function ProjectInCategoryDtoToJSON(json: any): ProjectInCategoryDto {
    return ProjectInCategoryDtoToJSONTyped(json, false);
}

export function ProjectInCategoryDtoToJSONTyped(value?: ProjectInCategoryDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'categoryId': value['categoryId'],
        'success': value['success'],
        'total': value['total'],
    };
}